refactor: move 'SuccessNotification' to '/notifications'
