+++ src/components/notifications/SuccessNotification.tsx
@@ -13,7 +13,7 @@ export default function SuccessNotification({
     <div
       className={`absolute top-0 right-0 mt-3 p-2 shadow bg-base-100 rounded-box w-52 transition transform ${opacity}`}
     >
-      The game was added succesfully!
+      The game was added successfully!
     </div>
   );
 }
