+++ src/routes/AddGame.tsx
@@ -6,11 +6,11 @@ import { createGame } from "../api/game/post/mutations";
 import { fetchPlayer } from "../api/player/get/queries";
 import { LoadingSpinner } from "../components/LoadingSpinner";
 import NavigationBar from "../components/NavigationBar";
+import SuccessNotification from "../components/notifications/SuccessNotification";
 import { PlayersSearchList } from "../components/PlayersSearchList";
 import { ScoreInput } from "../components/ScoreInput";
 import { SearchOpponentByName } from "../components/SearchOpponentByName";
 import { Steps } from "../components/Steps";
-import SuccessNotification from "../components/SuccessNotification";
 import { thirtyMinutes } from "../constants/time";
 import { useDebounce } from "../hooks/useDebounce";
 import { paths } from "../router/router";
